src from theme 03 Error Handling

+++ src/mobilele/src/main/java/bg/softuni/mobilele/model/entity/OfferEntity.java
@@ -56,7 +56,7 @@ public class OfferEntity extends BaseEntity {
     private String imageUrl;
 
     @Positive
-    private long mileage;
+    private int mileage;
 
     @NotNull
     @Column(nullable = false)
@@ -95,11 +95,11 @@ public class OfferEntity extends BaseEntity {
         return this;
     }
 
-    public long getMileage() {
+    public int getMileage() {
         return mileage;
     }
 
-    public OfferEntity setMileage(long mileage) {
+    public OfferEntity setMileage(int mileage) {
         this.mileage = mileage;
         return this;
     }
